refactor: domain 패키지 생성 및 클래스 파일 이동
