+++ src/test/java/lotto/LottoFactoryTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
