+++ src/test/java/lotto/LottoResultTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.LottoResult;
+import lotto.domain.Prize;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
